Preping Models and Data for GUI Implimentation
Collected data that will be used to build initial functionlaity in GUI. Located in gui_pickup/ directory.

Created script that is able to train and save linear regression model on selected data.

Will need to create similar scripts for KNN and SVR models so they can be picked up and used for implementing features in the GUI.